pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "comboi"
version = "0.1.0"
description = "Serverless DuckDB-based medallion ETL system"
authors = [{ name = "Your Name", email = "you@example.com" }]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "duckdb>=0.10.0",
    "pyyaml>=6.0.1",
    "rich>=13.7.0",
    "adlfs>=2024.4.0",
    "azure-identity>=1.15.0",
    "azure-keyvault-secrets>=4.8.0",
    "azure-storage-queue>=12.10.0",
    "azure-functions>=1.15.0",
    "azure-monitor-opentelemetry-exporter>=1.0.0",
    "splink>=3.10.3",
    "pandas>=2.2.0",
    "tenacity>=8.2.3",
    "typer>=0.12.3"
]

[project.scripts]
comboi = "comboi.cli:app"

